+++ src/main/java/de/johannzapf/bitcoin/terminal/util/Util.java
@@ -14,8 +14,8 @@ public class Util {
         return satoshi/100000000;
     }
 
-    public static int BTCToSatoshi(double btc){
-        return (int) Math.round(btc*100000000);
+    public static long BTCToSatoshi(double btc){
+        return Math.round(btc*100000000);
     }
 
     public static JSONObject parseJSON(HttpURLConnection conn) throws IOException {
